Refactor reminders to use a single layout for consistency and reduce duplication
