+++ src/client/modules/Reminders/OutstandingPropositionReminders.jsx
@@ -1,12 +1,7 @@
 import React from 'react'
 import { useLocation } from 'react-router-dom'
 import { connect } from 'react-redux'
-import {
-  SPACING,
-  FONT_SIZE,
-  HEADING_SIZES,
-  MEDIA_QUERIES,
-} from '@govuk-react/constants'
+import { SPACING, FONT_SIZE, HEADING_SIZES } from '@govuk-react/constants'
 import { Link } from 'govuk-react'
 import { BLACK, GREY_1, GREY_2 } from 'govuk-colours'
 import GridRow from '@govuk-react/grid-row'
@@ -22,12 +17,11 @@ import { format } from '../../utils/date'
 import { DATE_DAY_LONG_FORMAT } from '../../../common/constants'
 import { maxItemsToPaginate, itemsPerPage } from './constants'
 
-import { DefaultLayout, RoutedPagination } from '../../components'
+import { RoutedPagination } from '../../components'
 import CollectionHeader from './CollectionHeader'
-import RemindersMenu from './RemindersMenu'
+import RemindersLayout from './RemindersLayout'
 import Task from '../../components/Task'
 import urls from '../../../lib/urls'
-import Heading from './Heading'
 
 const List = styled('ol')({
   listStyleType: 'none',
@@ -72,37 +66,6 @@ const PaginationSummary = styled(Summary)({
   borderBottom: `solid 1px ${GREY_2}`,
 })
 
-const Container = styled('div')({
-  [MEDIA_QUERIES.DESKTOP]: {
-    display: 'flex',
-  },
-})
-
-const MenuContainer = styled('div')({
-  width: '100%',
-  [MEDIA_QUERIES.DESKTOP]: {
-    width: 'calc(33% - 20px)',
-    padding: SPACING.SCALE_2,
-  },
-})
-
-const ListContainer = styled('div')({
-  width: '100%',
-  [MEDIA_QUERIES.DESKTOP]: {
-    width: '67%',
-  },
-})
-
-const HomeLink = styled(Link)({
-  display: 'block',
-  marginTop: SPACING.SCALE_4,
-  marginBottom: SPACING.SCALE_4,
-  [MEDIA_QUERIES.DESKTOP]: {
-    marginLeft: 25,
-    marginBottom: 25,
-  },
-})
-
 const OutstandingPropositionReminders = ({
   outstandingPropositionsReminders,
 }) => {
@@ -117,73 +80,59 @@ const OutstandingPropositionReminders = ({
   )
 
   return (
-    <DefaultLayout
+    <RemindersLayout
       pageTitle={title}
-      heading={<Heading preHeading="Reminders for">{subject}</Heading>}
-      breadcrumbs={[{ link: urls.dashboard(), text: 'Home' }, { text: title }]}
+      subject={subject}
+      hasSettingsLink={false}
     >
-      <>
-        <Container>
-          <MenuContainer>
-            <RemindersMenu />
-          </MenuContainer>
-          <ListContainer>
-            <CollectionHeader settings={false} totalItems={count} />
-            {results.length === 0 ? (
-              <Summary data-test="no-reminders">You have no reminders.</Summary>
-            ) : (
-              <PaginationSummary data-test="pagination-summary">
-                {`Page ${page || 1} of ${totalPages}`}
-              </PaginationSummary>
-            )}
-            <Task.Status
-              name={TASK_GET_OUTSTANDING_PROPOSITIONS_REMINDERS}
-              id={ID}
-              startOnRender={{
-                payload: { page, sortby: qsParams.sortby },
-                onSuccessDispatch: REMINDERS__OUTSTANDING_PROPOSITIONS_LOADED,
-              }}
-            >
-              {() => (
-                <>
-                  <List data-test="reminders-list">
-                    {results.map(
-                      ({ id, name, deadline, investment_project }) => (
-                        <ListItem key={id} data-test="reminders-list-item">
-                          <GridRow>
-                            <GridCol>
-                              <ListItemHeader data-test="item-header">
-                                Due {format(deadline, DATE_DAY_LONG_FORMAT)}
-                              </ListItemHeader>
-                              <ListItemContent data-test="item-content">
-                                <Link
-                                  href={`${urls.investments.projects.propositions(
-                                    investment_project.id
-                                  )}`}
-                                >
-                                  {name}
-                                </Link>
-                              </ListItemContent>
-                              <ListItemFooter data-test="item-footer">
-                                Project code {investment_project.project_code}
-                              </ListItemFooter>
-                            </GridCol>
-                          </GridRow>
-                        </ListItem>
-                      )
-                    )}
-                  </List>
-                  <RoutedPagination initialPage={page} items={count || 0} />
-                </>
-              )}
-            </Task.Status>
-          </ListContainer>
-        </Container>
-        <HomeLink data-test="home-link" href={urls.dashboard()}>
-          Home
-        </HomeLink>
-      </>
-    </DefaultLayout>
+      <CollectionHeader settings={false} totalItems={count} />
+      {results.length === 0 ? (
+        <Summary data-test="no-reminders">You have no reminders.</Summary>
+      ) : (
+        <PaginationSummary data-test="pagination-summary">
+          {`Page ${page || 1} of ${totalPages}`}
+        </PaginationSummary>
+      )}
+      <Task.Status
+        name={TASK_GET_OUTSTANDING_PROPOSITIONS_REMINDERS}
+        id={ID}
+        startOnRender={{
+          payload: { page, sortby: qsParams.sortby },
+          onSuccessDispatch: REMINDERS__OUTSTANDING_PROPOSITIONS_LOADED,
+        }}
+      >
+        {() => (
+          <>
+            <List data-test="reminders-list">
+              {results.map(({ id, name, deadline, investment_project }) => (
+                <ListItem key={id} data-test="reminders-list-item">
+                  <GridRow>
+                    <GridCol>
+                      <ListItemHeader data-test="item-header">
+                        Due {format(deadline, DATE_DAY_LONG_FORMAT)}
+                      </ListItemHeader>
+                      <ListItemContent data-test="item-content">
+                        <Link
+                          href={`${urls.investments.projects.propositions(
+                            investment_project.id
+                          )}`}
+                        >
+                          {name}
+                        </Link>
+                      </ListItemContent>
+                      <ListItemFooter data-test="item-footer">
+                        Project code {investment_project.project_code}
+                      </ListItemFooter>
+                    </GridCol>
+                  </GridRow>
+                </ListItem>
+              ))}
+            </List>
+            <RoutedPagination initialPage={page} items={count || 0} />
+          </>
+        )}
+      </Task.Status>
+    </RemindersLayout>
   )
 }
